trekk ut TilpassKompetanserTilEndretUtbetalingAndelerService til egen fil
